requirements.txt
django
djangorestframework
psycopg2
djangorestframework_simplejwt
Pillow
django-cors-headers